Move PerFrame and PerSwapFrame into Vulkan dir and vk namespace

+++ include/BLIB/Render/Vulkan/AlignedBuffer.hpp
@@ -135,7 +135,7 @@ private:
 
     T* cast(std::size_t i);
     const T* cast(std::size_t i) const;
-    void ensureCapacity(std::uint32_t size);
+    void ensureCapacity(std::size_t size);
 };
 
 //////////////////////////// INLINE FUNCTIONS /////////////////////////////////
@@ -243,7 +243,7 @@ const T* AlignedBuffer<T>::cast(std::size_t i) const {
 }
 
 template<typename T>
-void AlignedBuffer<T>::ensureCapacity(std::uint32_t size) {
+void AlignedBuffer<T>::ensureCapacity(std::size_t size) {
     storage.resize(size * alignment);
 }
